requirements.txt
# Flask Framework
Flask==3.0.0
Flask-CORS==4.0.0

# LangChain and RAG Components
langchain==0.1.0
langchain-community==0.0.10
langchain-mongodb==0.1.3

# MongoDB
pymongo==4.6.1

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1

# Embeddings (local models)
sentence-transformers==2.7.0

# Environment and Configuration
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# Utilities
Werkzeug==3.0.1

# Additional dependencies
numpy>=1.26.0
torch>=2.0.0

